+++ lib/micro_serve.dart
@@ -1,8 +1,9 @@
 library;
 
 export 'src/service/server.dart';
-export 'src/constant/extension.dart';
+
 export 'src/constant/enum.dart';
+export 'src/constant/extension.dart';
 export 'src/helper/network.dart';
 export 'src/model/server_info_model.dart';
 export 'src/model/server_context_model.dart';