+++ lib/src/model/node_model.dart
@@ -1,4 +1,5 @@
-import 'package:micro_serve/src/common/common.dart';
+import 'package:micro_serve/src/constant/enum.dart';
+import 'package:micro_serve/src/model/server_context_model.dart';
 
 class Node {
   final Method method;